Feat: use tokens when communicating with backend in storage

+++ src/services/emergencyItemService.js
@@ -1,11 +1,30 @@
+import {useUserStore} from '@/stores/userStore';
+
 export function emergencyItemService() {
   const baseUrl = `${window.backendURL}/api/emergency/items`;
+  const userStore = useUserStore();
+
+  let token = userStore.token || sessionStorage.getItem("jwtToken") || "";
+
+  if (!token) {
+    throw new Error("No token found");
+  }
+
+  function getToken () {
+    return userStore.token || sessionStorage.getItem("jwtToken") || "";
+  }
 
   async function getEmergencyItems() {
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
     const response = await fetch(`${baseUrl}`, {
       method: 'GET',
       headers: {
         'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
       },
     });
     if (!response.ok) {
@@ -15,12 +34,19 @@ export function emergencyItemService() {
   }
 
   async function getEmergencyItemCategories() {
-    const response = await fetch(`${window.backendURL}/api/categories/categories`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
+    const response = await fetch(
+        `${window.backendURL}/api/categories/categories`, {
+          method: 'GET',
+          headers: {
+            'Content-Type': 'application/json',
+            'Authorization': `Bearer ${currentToken}`,
+          },
+        });
     if (!response.ok) {
       throw new Error('Failed to fetch emergency item categories');
     }
@@ -28,10 +54,16 @@ export function emergencyItemService() {
   }
 
   async function getEmergencyItemUnits() {
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
     const response = await fetch(`${window.backendURL}/api/units`, {
       method: 'GET',
       headers: {
         'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
       },
     });
     if (!response.ok) {
@@ -41,7 +73,18 @@ export function emergencyItemService() {
   }
 
   async function getEmergencyItemByCategoryId(categoryId) {
-    const response = await fetch(`${baseUrl}/categories/${categoryId}`);
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
+    const response = await fetch(`${baseUrl}/categories/${categoryId}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
+      }
+    });
     if (!response.ok) {
       throw new Error('Failed to fetch emergency items by category');
     }
@@ -49,7 +92,18 @@ export function emergencyItemService() {
   }
 
   async function getEmergencyItemById(id) {
-    const response = await fetch(`${baseUrl}/${id}`);
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
+    const response = await fetch(`${baseUrl}/${id}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
+      }
+    });
     if (!response.ok) {
       throw new Error('Failed to fetch emergency item by ID');
     }
@@ -57,10 +111,16 @@ export function emergencyItemService() {
   }
 
   async function createEmergencyItem(item) {
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
     const response = await fetch(baseUrl, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
       },
       body: JSON.stringify(item),
     });
@@ -71,10 +131,16 @@ export function emergencyItemService() {
   }
 
   async function updateEmergencyItem(item) {
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
     const response = await fetch(`${baseUrl}`, {
       method: 'PUT',
       headers: {
         'Content-Type': 'application/json',
+        'Authorization': `Bearer ${currentToken}`,
       },
       body: JSON.stringify(item),
     });
@@ -85,8 +151,16 @@ export function emergencyItemService() {
   }
 
   async function deleteEmergencyItem(id) {
+    const currentToken = getToken();
+    if (!currentToken) {
+      throw new Error("No token found");
+    }
+
     const response = await fetch(`${baseUrl}/${id}`, {
       method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${currentToken}`,
+      }
     });
     if (!response.ok) {
       throw new Error('Failed to delete emergency item');
